Easier to maintain pytest
